json parse fixes, track change's working

+++ app/src/main/java/com/dinaraparanid/prima/utils/web/genius/songs_response/Author.kt
@@ -7,7 +7,7 @@ import java.io.Serializable
 data class Author(
     @Expose
     @JvmField
-    val attribution: Int,
+    val attribution: Double,
 
     @Expose
     @JvmField
